% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsqLoad.R
\name{dsqLoad}
\alias{dsqLoad}
\title{Load the result of an OMOP query in the remobe session(s)}
\usage{
dsqLoad(
  symbol = NULL,
  domain = NULL,
  query_name = NULL,
  where_clause = NULL,
  row_limit = NULL,
  db_connection = NULL,
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{symbol}{a character, the name of the dataframe where the resultset will be loaded. It defaults to the query_name.}

\item{domain}{a character, the query domain (ex. 'care_site')}

\item{query_name}{the query name as it appears in the result of dsqShowQueries}

\item{where_clause}{an optional where clause that can be appended to the query (without the 'where' keyword)}

\item{row_limit}{maximum number of row retrieved}

\item{db_connection}{a character, the name of the connection to the database. This must exist in the remote session(s) - it can be created with datashield.assign.resource().
If no db_connection is provided, the first connection found in the renote session (alphabetically) will be used.}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\value{
the query result
}
\description{
Load remotely one of the preset queries (previously retrieved with dsqShowQueries)
}
