% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsqRun.R
\name{dsqRun}
\alias{dsqRun}
\title{Run a OMOP dictionary aggregate query}
\usage{
dsqRun(
  domain = NULL,
  query_name = NULL,
  input = NULL,
  db_connection = NULL,
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{domain}{a character, the query domain (ex. 'care_site')}

\item{query_name}{the query name as it appears in the result of dsqShowQueries}

\item{input}{a vector of input parameters, in the same order as they appear in the text of the query. Information is available in dsqShowQueries() for each case.}

\item{db_connection}{a character, the name of the connection to the database. This must exist in the remote session(s) - it can be created with datashield.assign.resource().
If no db_connection is provided, the first connection found in the renote session (alphabetically) will be used.}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\value{
the query result
}
\description{
Run one of the preset queries (previously retrieved with dsqShowQueries)
}
