% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsqShowQueries.R
\name{dsqShowQueries}
\alias{dsqShowQueries}
\title{Show all available queries}
\usage{
dsqShowQueries(async = TRUE, datasources = NULL)
}
\arguments{
\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\value{
a list containing all the available queries with documentation
}
\description{
Return all the available queries explained and grouped by domain.
}
