% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsqLoad.R
\name{dsqLoad}
\alias{dsqLoad}
\title{Load the result of an OMOP query in the remote session(s)}
\usage{
dsqLoad(
  symbol = NULL,
  domain = NULL,
  query_name = NULL,
  input = NULL,
  where_clause = NULL,
  row_limit = NULL,
  row_offset = 0,
  db_connections = NULL,
  union = TRUE,
  exclude_cols_regex = c("source"),
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{symbol}{a character, the name of the data frame where the resultset will be loaded. It defaults to the query_name.}

\item{domain}{a character, the query domain (ex. 'care_site')}

\item{query_name}{the query name as it appears in the result of dsqShowQueries}

\item{input}{a vector of input parameters, in the same order as they appear in the text of the query. Information is available in dsqShowQueries() for each case.}

\item{where_clause}{an optional where clause that can be appended to the query (without the 'where' keyword)}

\item{row_limit}{maximum number of row retrieved}

\item{db_connections}{a vector, the name of the connection(s) to the database. They must exist in the remote session(s) - it can be created with datashield.assign.resource().
If no db_connection is provided, all the connections found in the renote session will be used. One more column called "database"
will be added to the query results and it will be populated with the respective connection names. The next argument (union)
will goevern further behaviour.}

\item{union}{a logical, if TRUE (the default), the result sets across the multiple connections will be concatenated in one single data frame.
If FALSE, one data frame will be created for each connection (with the connection name concatenated to the original name)}

\item{exclude_cols_regex, }{a vector of regular expressions used to filter out certain columns from the result. By default it filters out the columns containting 'source' in their name.
If FALSE, one data frame will be created for each connection (with the connection name concatenated to the original name)}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\value{
the query result
}
\description{
Load remotely one of the preset queries (previously retrieved with dsqShowQueries)
}
